ENG-122 Eliminated RuntimeConfigurationImpl class
