+++ src/main/java/org/ambraproject/rhino/config/RuntimeConfiguration.java
@@ -23,40 +23,92 @@
 package org.ambraproject.rhino.config;
 
 import java.net.URI;
+import com.google.common.base.Preconditions;
 
-/**
- * Interface that represents configurable values that are only known at server startup time.
- */
-public interface RuntimeConfiguration {
+public class RuntimeConfiguration {
+  private boolean prettyPrintJson = false;
+  private URI contentRepoUrl;
+  private String editorialBucket;
+  private String corpusBucket;
+  private URI taxonomyUrl;
+  private String thesaurus;
 
   /**
    * determine if json should be formatted or not
    *
    * @return false when formatting is turned off
    */
-  public boolean getPrettyPrintJson();
+  public boolean getPrettyPrintJson() {
+    return this.prettyPrintJson;
+  }
+
+  public void setPrettyPrintJson(boolean prettyPrintJson) {
+    this.prettyPrintJson = prettyPrintJson;
+  }
 
   /**
    * @return the URI of the content repo server
    */
-  public URI getContentRepoUrl();
+  public URI getContentRepoUrl() {
+    return this.contentRepoUrl;
+  }
+
+  public void setContentRepoUrl(URI contentRepoUrl) {
+    Preconditions.checkNotNull(contentRepoUrl,
+        "CONTENT_REPO_URL is required and must be a valid URL");
+    Preconditions.checkState(contentRepoUrl.isAbsolute(),
+        "CONTENT_REPO_URL is required and must be a valid URL");
+    this.contentRepoUrl = contentRepoUrl;
+  }
 
   /**
-   * Return the content repository bucket for the corpus of articles. The application will write to this bucket when
-   * ingesting articles and read from it when serving article assets.
+   * Return the content repository bucket for the corpus of articles. The application will write to
+   * this bucket when ingesting articles and read from it when serving article assets.
    *
    * @return the corpus bucket name
    */
-  public String getCorpusBucket();
-  
-  /**
-   * Return the content repository bucket from which the system should pick up editorial (non-article) content. 
+  public String getCorpusBucket() {
+    return this.corpusBucket;
+  }
+
+  public void setCorpusBucket(String corpusBucket) {
+    Preconditions.checkNotNull(corpusBucket, "CORPUS_BUCKET is required");
+    Preconditions.checkState(!corpusBucket.equals(""), "CORPUS_BUCKET is required");
+    this.corpusBucket = corpusBucket;
+  }
+
+  /*
+   * Return the content repository bucket from which the system should pick up editorial
+   * (non-article) content.
+   * 
    * @return the homepage bucket name
-   *
    */
-  public String getEditorialBucket();
+  public String getEditorialBucket() {
+    return this.editorialBucket;
+  }
 
-  public URI getTaxonomyUrl();
+  public void setEditorialBucket(String editorialBucket) {
+    this.editorialBucket = editorialBucket;
+  }
 
-  public String getThesaurus();
+  public URI getTaxonomyUrl() {
+    return this.taxonomyUrl;
+  }
+
+  public void setTaxonomyUrl(URI taxonomyUrl) {
+    Preconditions.checkNotNull(taxonomyUrl, "TAXONOMY_URL is required and must be a valid URL");
+    Preconditions.checkState(taxonomyUrl.isAbsolute(),
+        "TAXONOMY_URL is required and must be a valid URL");
+    this.taxonomyUrl = taxonomyUrl;
+  }
+
+  public String getThesaurus() {
+    return this.thesaurus;
+  }
+
+  public void setThesaurus(String thesaurus) {
+    Preconditions.checkNotNull(thesaurus, "THESAURUS is required");
+    Preconditions.checkState(!thesaurus.equals(""), "THESAURUS is required");
+    this.thesaurus = thesaurus;
+  }
 }
